+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.9.0 2026-07-11
 ### Breaking change
 - Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Configuration#how-to-delete-and-re-add-the-solarcharger-integration).
+- Renamed charge sub-state from "Self-paused" to "Self-derated".
+- Renamed counter "Self-paused today" to "Self-derated today" to support device that can vary current by itself.
 - Replaced "Start on pause" switch with "Start state" selector.
 - Replaced "End on max consumed energy" switch with a more general "End on condition" switch.
 - Removed "Max consumed energy limit" number.
@@ -15,6 +17,7 @@
 - Added "[Start state](https://github.com/flashg1/SolarCharger/wiki/User-guide#start-state)" selector for Auto, Charge or Pause start states.
 - Added "[Exit condition sensor](https://github.com/flashg1/SolarCharger/wiki/User-guide#exit-condition-sensor)" selector and "End on condition" switch.
 ### Improvement
+- Support power allocation to device that can vary current by itself.
 - Always read current before setting new current in case device current is out of sync with SolarCharger.
 - Refactored select.py to pass in available options as a class parameter.